// src/client/themes/vizhubTheme/vizhubThemeColors.ts
import { vizhubThemeColorsComputation } from './vizhubThemeColorsComputation';

type ColorSettings = {
  color?: string;
  backgroundColor?: string;
  cursor?: string;
  fontWeight?: string;
  textDecoration?: string;
  borderBottom?: string;
  border?: string;
};

type VizHubColors = {
  container: ColorSettings;
  headerBackgroundColor: string;
  selectionBackground: string;
  lineNumbers: ColorSettings;
  comment: ColorSettings;
  meta: ColorSettings;
  fatCursor: ColorSettings;
  default: ColorSettings;
  link: ColorSettings;
  searching: ColorSettings;
  keyword: ColorSettings;
  tag: ColorSettings;
  variable: ColorSettings;
  variable2: ColorSettings;
  variable3: ColorSettings;
  definition: ColorSettings;
  qualifier: ColorSettings;
  builtin: ColorSettings;
  header: ColorSettings;
  string: ColorSettings;
  string2: ColorSettings;
  atom: ColorSettings;
  number: ColorSettings;
  attribute: ColorSettings;
  property: ColorSettings;
  colors: string[];
  operator: ColorSettings;
  caretColor: string;
  gutter: ColorSettings;
  gutters: ColorSettings;
  matchingBracket: ColorSettings;
};

// Generated by vizhubThemeColorsComputation in vizhubTheme.test
export const vizhubThemeColors: VizHubColors =
  vizhubThemeColorsComputation(0);